FIX - corrigir erros encontrados no commit anterior
