+++ src/model/Categoria.java
@@ -1,0 +1,14 @@
+
+package model;
+
+public class Categoria {
+    private String nome;
+
+    public String getNome() {
+        return nome;
+    }
+
+    public void setNome(String nome) {
+        this.nome = nome;
+    }
+}
