--- src/model/CategoriasClass.java
@@ -1,23 +1,0 @@
-
-package Classes;
-
-public class CategoriasClass {
-    private int id;
-    private String nome;
-
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    public String getNome() {
-        return nome;
-    }
-
-    public void setNome(String nome) {
-        this.nome = nome;
-    }
-}
